+++ src/core/server/getters/players.ts
@@ -1,6 +1,7 @@
 import * as alt from 'alt-server';
 import * as Athena from '@AthenaServer/api';
 import * as vehicle from './vehicle';
+import { Character } from '@AthenaShared/interfaces/character';
 
 /**
  * Return all players currently online and logged into a character.
@@ -18,6 +19,12 @@ export function online(): alt.Player[] {
     });
 }
 
+/**
+ * Return all players with weapons out.
+ *
+ * @export
+ * @return {alt.Player[]}
+ */
 export function onlineWithWeapons(): alt.Player[] {
     return [...alt.Player.all].filter((p) => {
         if (!p.valid) {
@@ -36,6 +43,7 @@ export function onlineWithWeapons(): alt.Player[] {
         return true;
     });
 }
+
 /**
  * Creates an array of players who are closest to a position.
  * Array is automatically sorted into ascending order.
@@ -189,4 +197,15 @@ export function drivingSpecificModel(model: string | number): alt.Player[] {
  */
 export function inVehicle(vehicle: alt.Vehicle): alt.Player[] {
     return alt.Player.all.filter((x) => x.vehicle && x.vehicle.id === vehicle.id);
+}
+
+/**
+ * Get all online players as characters.
+ *
+ * @export
+ * @template T
+ * @return {*}
+ */
+export function asCharacters<T = Character>() {
+    return Athena.document.character.getAllOnline<T>();
 }